Add number funcs param & assignment -> defval
In the `SyntheticCaseA` script, we have the following code:

````
if numfun(h) > 0
              % Estimate the total mass change
              % How is numfun(h) supposed to be involved here?
              [ESTsignal,ESTresid,ftests,extravalues,total,...
               alphavarall,totalparams,totalparamerrors,totalfit,...
               functionintegrals,alphavar] = slept2resid(slept,thedates,...
                [1 365.0],[],[],CC,TH,numfun(h));
               
              allslopes{h}(counter) = totalparams(2)*365;
           else
# ...
````

This code presumes the existence of a parameter like the one I am proposing, and does not work without it.  I don't think my proposed change will break anything (fingers crossed), and if I understand the code in `SyntheticCaseA` correctly, I believe it will help make that code work on regions such as Iceland where the Shannon number is very small.

The alternative would be to round up the Shannon number by default, but I worry that this might have unintended consequences elsewhere in scripts which call this script already.