Updated glm2plmt function for RL06
Updated glm2plmt function for RL06 which includes the C(2,0) coefficient for RL06. The degree 1 coefficient is not released yet for RL06 and thus the degree 1 coefficient data used is from the latest update of RL05. This update includes a description of the RL06 naming convention for the gzip files.